\name{Clinical.data}
\alias{Clinical.data}
\docType{data}
\title{
    A data frame for example clinical data
}
\description{
    A data frame consisting of survival status and time for each sample
}
\usage{data(sampleclinical)}
\format{
    A data frame with survival information and times on the 50 samples
}
\value{
    A data frame with survival information and times on the 50 samples
}
\source{
    The data was provided from \pkg{IMAS}
}
\examples{
    data(sampleclinical)
}
\keyword{datasets}