\name{GroupSam}
\alias{GroupSam}
\docType{data}
\title{
    Group of each sample.
}
\description{
    A list object of a group of each sample
}
\usage{data(sampleGroups)}
\format{
    A list object of a group information on the 50 individuals
}
\value{
    A list object of a group information on the 50 individuals
}
\source{
    The data was provided from \pkg{IMAS}
}
\examples{
    data(sampleGroups)
}
\keyword{datasets}