\name{IMAS-package}
\alias{IMAS-package}
\alias{IMAS}
\docType{package}
\title{\pkg{IMAS}: Integrative analysis of Multi-omics data for Alternative Splicing}
\description{
\pkg{IMAS} offers two components. First, \code{\link{RatioFromReads}} estimates PSI values of a given alternatively spliced exon using both of paired-end and junction reads. See the examples at \code{\link{RatioFromReads}}. Second, \code{\link{CompGroupAlt}}, \code{\link{MEsQTLFinder}}, and \code{\link{ClinicAnalysis}} can be used for further analysis using estimated PSI values. We described more detailed information on usage at the package vignette.
}
\author{
    Seonggyun Han, Younghee Lee
}
\keyword{package}
