\name{sampleMedata}
\alias{sampleMedata}
\docType{data}
\title{
    Methylation level data
}
\description{
    Methylation level of 5 loci (beta value), which are located in the PRGA gene for 50 samples. We generated a simulation data set of methylation level (beta value) for each locus, that significantly differs between two groups (PR-positive and PR-negative), while other methylation loci are not different. The detailed overview of the data is described in the vignette.
}
\usage{
    data(samplemethyl)
}
\format{
    A data frame with levels of 5 methylation locus on the 50 samples
}
\value{
    A data frame with levels of 5 methylation locus on the 50 samples
}
\examples{
    data(samplemethyl)
}
\keyword{datasets}