\name{sampleMelocus}
\alias{sampleMelocus}
\docType{data}
\title{
    Genomic locus of methylations
}
\description{
    Genomic location of 5 methylation loci located in the PRGA gene for 50 samples, which are matched with methylation level data provided in \pkg{IMAS}. The detailed overview of the data is described in the vignette.
}
\usage{data(samplemethyllocus)}
\format{
    A data frame with genomic locus of 5 methylations
}
\value{
    A data frame with genomic locus of 5 methylations
}
\examples{
    data(samplemethyllocus)
}
\keyword{datasets}
