\name{samplesnp}
\alias{samplesnp}
\docType{data}
\title{
    Genotype data
}
\description{
    Genotype data of five SNPs located in the PRGA gene for 50 samples (half of whom are assigned as PR-positive and the other half PR-negative), which is used in analysis with \pkg{IMAS}. We generated a simulation data set of genotypes for each SNP.  Among five SNPs, three are associated with PSI levels for 50 samples, while two SNPs are not. The detailed overview of the data is described in the vignette.
}
\usage{data(samplesnp)}
\format{
    A data frame with genotypes of 5 SNPs on the 50 samples
}
\value{
    A data frame with genotypes of 5 SNPs on the 50 samples
}
\source{
    The data was provided from \pkg{IMAS}
}
\examples{
    data(samplesnp)
}
\keyword{datasets}