\name{samplesnplocus}
\alias{samplesnplocus}
\docType{data}
\title{
    Genomic locus of SNPs
}
\description{
    Genomic locus of five SNPs located in the PRGA gene for 50 samples, which are matched with SNP genotype data provided in \pkg{IMAS}. The detailed overview of the data is described in the vignette.
}
\usage{data(samplesnplocus)}
\format{
    A data frame with genomic locus of 5 SNPs
}
\value{
    A data frame with genomic locus of 5 SNPs
}
\examples{
    data(samplesnplocus)
}
\keyword{datasets}