\name{Clinical.data}
\alias{Clinical.data}
\docType{data}
\title{
    A data frame for clinical data
}
\description{
    A data frame including survival status and time for each sample. This data is a simulated clinical data for 50 samples (half of whom are assigned as PR-positive and the other half PR-negative), which is used in analysis with \pkg{IMAS}. The detailed overview of the data is described in the vignette.
}
\usage{data(sampleclinical)}
\format{
    A data frame with survival information and times on the 50 samples
}
\value{
    A data frame with survival information and times on the 50 samples
}
\keyword{datasets}