\name{GroupSam}
\alias{GroupSam}
\docType{data}
\title{
    Group of each sample.
}
\description{
    A list object comprising sample names belonging to each group, PR-positive and PR-negative. This data is a simulated clinical data for 50 samples (half of whom are assigned as PR-positive and the other half PR-negative). The detailed overview of the data is described in the vignette.
}
\usage{data(sampleGroups)}
\format{
    A list object including a group information on the 50 samples
}
\value{
    A list object including a group information on the 50 samples
}
\examples{
    data(sampleGroups)
}
\keyword{datasets}