\name{samplebamfiles}
\alias{samplebamfiles}
\docType{data}
\title{
    A data frame for example expression bam files.
    }
\description{
    A path and identifier of bam files including 50 samples
    }
\usage{data(bamfilestest)}
\format{
    A data frame with paths and identifiers on the 50 samples
    }
\value{
    A data frame with paths and identifiers on the 50 samples
    }
\source{
    The data was provided from \pkg{IMAS}
    }
\examples{
    data(bamfilestest)
    }
\keyword{datasets}