\name{sampleMedata}
\alias{sampleMedata}
\docType{data}
\title{
    Methylation levels data
}
\description{
    Methylation levels data including 50 individuals
}
\usage{data(samplemethyl)}
\format{
    A data frame with levels of 5 methylation locus on the 50 individuals
}
\value{
    A data frame with levels of 5 methylation locus on the 50 individuals
}
\source{
    The data was provided from \pkg{IMAS}
}
\examples{
    data(samplemethyl)
}
\keyword{datasets}