\name{sampleMelocus}
\alias{sampleMelocus}
\docType{data}
\title{
    Methylation locus
}
\description{
    Methylation locus
}
\usage{data(samplemethyllocus)}
\format{
    A data frame with 5 methylations and locus of them
}
\value{
    A data frame with 5 methylations and locus of them
}
\source{
    The data was provided from \pkg{IMAS}
}
\examples{
    data(samplemethyllocus)
}
\keyword{datasets}
