\name{samplesnp}
\alias{samplesnp}
\docType{data}
\title{
    genotype data
}
\description{
    genotype data
}
\usage{data(samplesnp)}
\format{
    A data frame with genotypes of 5 SNPs on the 50 individuals
}
\value{
    A data frame with genotypes of 5 SNPs on the 50 individuals
}
\source{
    The data was provided from \pkg{IMAS}
}
\examples{
    data(samplesnp)
}
\keyword{datasets}