\name{samplesnplocus}
\alias{samplesnplocus}
\docType{data}
\title{
    SNP locus
}
\description{
    SNP locus
}
\usage{data(samplesnplocus)}
\format{
    A data frame with 5 SNPs and locus of them
}
\value{
    A data frame with 5 SNPs and locus of them
}
\source{
    The data was provided from \pkg{IMAS}
}
\examples{
    data(samplesnplocus)
}
\keyword{datasets}